second-approach/scripts/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-pinecone>=0.0.1
langchain-community>=0.0.10
pinecone-client>=3.0.0
openai>=1.12.0
python-dotenv>=1.0.0
pdf2image>=1.16.3
pytesseract>=0.3.10
tqdm>=4.66.1 